fix floats.zig

+++ ziglings/060_floats.zig
@@ -29,8 +29,6 @@
 //     var foo: f16 = 13.5 * 5;   // ERROR!
 //     var foo: f16 = 13.5 * 5.0; // No problem, both are floats
 //
-// Please fix the two float problems with this program and
-// display the result as a whole number.
 
 const print = @import("std").debug.print;
 
@@ -40,7 +38,7 @@ pub fn main() void {
     //
     // We'll convert this weight from tons to kilograms at a
     // conversion of 907.18kg to the ton.
-    var shuttle_weight: f16 = 907.18 * 2200;
+    var shuttle_weight: f32 = 907.18 * 2200.0;
 
     // By default, float values are formatted in scientific
     // notation. Try experimenting with '{d}' and '{d:.3}' to see
